Refactor document generation test values
